pyproject.toml
[project]
name = "ethos"
version = "0.1.0"
description = "EthosNews — participatory, intent-driven news ecosystem"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # ── Web framework ──────────────────────────────────────────────────────
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",

    # ── Config & validation ────────────────────────────────────────────────
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # ── Database ───────────────────────────────────────────────────────────
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",

    # ── Cache ──────────────────────────────────────────────────────────────
    "redis>=5.0.0",

    # ── Vector DB ──────────────────────────────────────────────────────────
    "qdrant-client>=1.9.0",

    # ── LLM providers ─────────────────────────────────────────────────────
    "groq>=0.9.0",
    "openai>=1.30.0",
    "google-generativeai>=0.7.0",

    # ── Embeddings & NLP ──────────────────────────────────────────────────
    "sentence-transformers>=3.0.0",
    "spacy>=3.7.0",

    # ── Clustering ────────────────────────────────────────────────────────
    "hdbscan>=0.8.33",
    "numpy>=1.26.0",

    # ── Ingestion ─────────────────────────────────────────────────────────
    "gdelt>=0.1.10",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
